+++ src/pages/index.js
@@ -3,7 +3,7 @@ import Img from 'gatsby-image'
 import React from 'react'
 import SEO from 'react-seo-component'
 import styled from 'styled-components'
-import { Layout } from '../components/Layout'
+import { Layout } from '../components/layout'
 import { useSiteMetadata } from '../hooks/use-site-metadata'
 
 const IndexWrapper = styled.main``
