+++ src/templates/blog-post-template.js
@@ -2,7 +2,7 @@ import { graphql, Link } from 'gatsby'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
 import React from 'react'
 import SEO from 'react-seo-component'
-import { Layout } from '../components/Layout'
+import { Layout } from '../components/layout'
 import { useSiteMetadata } from '../hooks/use-site-metadata'
 
 export default ({ data, pageContext }) => {
